+++ src/core/ChatSDKError.ts
@@ -66,7 +66,12 @@ export enum ChatSDKErrorName {
     AuthenticatedUserTokenNotFound = "AuthenticatedUserTokenNotFound",
     /** Failure in mid-conversation authentication */
     MidConversationAuthFailure = "MidConversationAuthFailure",
-
+    /** Failure in sending read receipt */
+    SendReadReceiptFailure = "SendReadReceiptFailure",
+    /** Invalid parameters for send read receipt (missing messageId) */
+    SendReadReceiptInvalidParams = "SendReadReceiptInvalidParams",
+    /** Failure in retrieving unread message count */
+    UnreadMessageCountRetrievalFailure = "UnreadMessageCountRetrievalFailure",
 }
 
 export class ChatSDKError {